+++ lib/flutter_ddi.dart
@@ -3,6 +3,10 @@ library flutter_ddi;
 export 'package:dart_ddi/dart_ddi.dart';
 
 export 'src/extensions/flutter_ddi_context.dart';
+export 'src/extensions/flutter_ddi_router_extension.dart';
+export 'src/features/interceptors/flutter_ddi_interceptor.dart';
+export 'src/features/interceptors/interceptor_result.dart';
+export 'src/features/interceptors/module_interceptor.dart';
 export 'src/features/listener/flutter_ddi_event_listener.dart';
 export 'src/features/listener/flutter_ddi_stream_listener.dart';
 export 'src/features/state/event_listener_state.dart';
@@ -10,5 +14,4 @@ export 'src/features/state/stream_listener_state.dart';
 export 'src/interfaces/error_module_interface.dart';
 export 'src/interfaces/loader_module_interface.dart';
 export 'src/modules/flutter_ddi_modules_define.dart';
-export 'src/route/flutter_ddi_router.dart';
 export 'src/wigets/flutter_ddi_builder.dart';